prep to write main data parquet file
